// src/components/chat/types.ts
export interface ChatPreview {
  id: string;
  name: string;
  lastMessage: string;
  timestamp: Date | string;
  unread: number;
  photoURL: string;
  status?: string;
  lastSeen?: Date | string;
  pinned?: boolean;
  isAnonymous?: boolean;
  isGroup?: boolean;
  type?: 'direct' | 'group' | 'service';
}

export interface Contact {
  id: string;
  name: string;
  status: 'online' | 'offline' | 'unregistered';
  lastSeen?: string;
  photoURL?: string;
  phoneNumber?: string;
  isAnonymous?: boolean;
}